Moved representations in own directory for better orientation.
