+++ ftw/zipexport/representations/archetypes.py
@@ -1,25 +1,11 @@
 from ftw.zipexport.interfaces import IZipRepresentation
+from ftw.zipexport.representations.general import NullZipRepresentation
 from Products.CMFCore.interfaces import IFolderish
 from zope.component import adapts
 from zope.component import getMultiAdapter
 from zope.interface import implements
 from zope.interface import Interface
 from Products.ATContentTypes.interfaces.file import IFileContent
-
-
-class NullZipRepresentation(object):
-    """
-    defualt zip export implementation
-    """
-    implements(IZipRepresentation)
-    adapts(Interface, Interface)
-
-    def __init__(self, context, request):
-        self.context = context
-        self.request = request
-
-    def get_files(self, path_prefix="", recursive=True, toplevel=True):
-        return
 
 
 class FolderZipRepresentation(NullZipRepresentation):
