+++ ftw/zipexport/representations/general.py
@@ -1,0 +1,19 @@
+from ftw.zipexport.interfaces import IZipRepresentation
+from zope.component import adapts
+from zope.interface import implements
+from zope.interface import Interface
+
+
+class NullZipRepresentation(object):
+    """
+    defualt zip export implementation
+    """
+    implements(IZipRepresentation)
+    adapts(Interface, Interface)
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def get_files(self, path_prefix="", recursive=True, toplevel=True):
+        return
